JDP191102-25 Implemented methods in UserController. Added UserToken class to generate random Token
