--- src/main/java/com/kodilla/ecommercee/Cart.java
@@ -1,20 +1,0 @@
-package com.kodilla.ecommercee;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class Cart {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-}
