--- src/main/java/com/kodilla/ecommercee/UserController.java
@@ -1,37 +1,0 @@
-package com.kodilla.ecommercee;
-
-import java.security.SecureRandom;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-
-@RestController
-@RequestMapping("/v1/user")
-public class UserController {
-
-    @Autowired
-    private UserService service;
-
-    @Autowired
-    private UserMapper mapper;
-
-    @PostMapping(value = "create", consumes = APPLICATION_JSON_VALUE)
-    public void createUser(@RequestBody UserDto userDto) {
-        service.saveUser(mapper.mapToUser(userDto));
-    }
-
-    @PutMapping(value = "block")
-    public void blockUser(Long userId) {
-
-    }
-
-    @PutMapping(value = "generateToken")
-    public String generateToken(Long userId) {
-        SecureRandom random = new SecureRandom();
-        byte bytes[] = new byte[20];
-        random.nextBytes(bytes);
-        String token = bytes.toString();
-        return token;
-    }
-}
