+++ src/main/java/com/kodilla/ecommercee/domain/UserDto.java
@@ -1,10 +1,12 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.Setter;
 
 @AllArgsConstructor
 @Getter
+@Setter
 public class UserDto {
 
     private Long id;
