+++ src/main/java/com/kodilla/ecommercee/mapper/UserMapper.java
@@ -1,5 +1,7 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.mapper;
 
+import com.kodilla.ecommercee.domain.User;
+import com.kodilla.ecommercee.domain.UserDto;
 import org.springframework.stereotype.Component;
 
 @Component
