+++ src/main/java/com/kodilla/ecommercee/service/UserService.java
@@ -1,5 +1,8 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.service;
 
+import com.kodilla.ecommercee.domain.User;
+import com.kodilla.ecommercee.repository.UserRepository;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,5 +14,9 @@ public class UserService {
 
     public User saveUser(User user) {
         return repository.save(user);
+    }
+
+    public Optional<User> findUser(Long userId) {
+        return repository.findById(userId);
     }
 }
